+++ docs/CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## 15.0.0
 
+### 15.0.0 Breaking Changes
+
+- `[Icons]` The `SohoIconsEmptyComponent` can now be configured to use one of `"colorful"`, `"classic"` or `"new"` icons. By default, the newly added `"colorful"` icons will be used. To revert to the classic icons, applications must explicitly configure this:
+
+```html
+<soho-icons-empty icons="classic"></soho-icons-empty>
+```
+
 ### 15.0.0 Features
 
 - `[Datagrid]` Added a setting to set the color of the header in datagrid (`dark` or `light`). ([#7008](https://github.com/infor-design/enterprise/issues/7008)) `EA`
 - `[General]` Updated to NG 15 d made small fixes. See the `UPGRADING.md` guide for details. ([#1316](https://github.com/infor-design/enterprise-ng/issues/1316))
+- `[Icons]` Merge old, new and "colorful" empty state icons into `SohoIconsEmptyComponent`, with the ability to force the use of colorful icons. ([#1418](https://github.com/infor-design/enterprise-ng/issues/1418))
 
 ### 15.0.0 Fixes